+++ internal/proxypool/test_test.go
@@ -1,0 +1,10 @@
+package proxypool
+
+import "testing"
+
+func TestTestProxyInvalidURL(t *testing.T) {
+	res := TestProxy("not-a-valid-url", "http", "")
+	if res.OK {
+		t.Errorf("TestProxy returned OK for an invalid URL")
+	}
+}